% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2ctmeval.emepdata.R
\name{convert2ctmeval.emepdata}
\alias{convert2ctmeval.emepdata}
\title{converts an emepdata variable into a variables usable in the ctmeval package}
\format{R Code}
\usage{
convert2ctmeval.emepdata(emepdata, vlevels = 0, calc_mean = TRUE,
  mean_type = "daily", dsum = FALSE)
}
\arguments{
\item{emepdata}{(emepdata): list in the emepdata format}

\item{vlevels}{(integer array): array containing valid flags (all values with flags that
are not in this array are set to NA)}

\item{calc_mean}{(logical): calculate mean values if TRUE, standard is TRUE}

\item{mean_type}{(character): calculate which type of mean? standard is daily}

\item{dsum}{(logical): calculate sum instead of mean if TRUE, standard is FALSE}
}
\value{
timeseries of observations in observationsts format (in accordance
                    with the ctmeval package v0.9.5.1)
}
\description{
converts a list in emepdata format into a list of the format
produced by ctmeval (observationts format). However, the returned
list contains three additional elements that are
     $station$altitude
     $timeser$hour
     $method_ref
}
\examples{
# read emep data file
myEmepdata = getEmptyEmepData() # TODO!!!
# convert into ctmeval format
data_ctmeval = convert2ctmeval.emepdata(myEmepdata)

Find out if sum instead of mean should be applied. Further below, 'fun_apply'
is the 'FUN' input argument of an 'apply' call which calculates the mean/sum.
'ctmeval_obs' will be the return variable in the end of this function. It is a
list and has one list element per measured species. Each list element will have
the structure of 'dummy_obs'.
'dummy_obs' has the structure of one 'observationts' list which is return from
a 'getobservationts' call of the 'ctmeval' package v 0.9.5.1. It has three 
additional elements which are
   $station$altitude,
   $timeser$hour, and
   $method_ref.
'dummy_obs' is the template for all elements in ctmeval_obs (copied for each 
element) and is filled with all information that is equal for all measurements.
These information are '$station' and '$type'. All other list members are created
but set to NULL. 
iterate all measured species/variables
copy dummy_obs and find out which species we are copying
copy meta data
copy times and values
remove invalid values
do mean/sum calculations
daily mean/sum values
We only have to do something if the data is not already in daily
resolution (emepdata$addDataInfo$`Resolution code` == '1d').
mean_type == emepdata$addDataInfo$`Resolution code`
else ...
add new observationts list to the return-list
return the return-list
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

