% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDuplicates.emepdata.R
\name{findDuplicates.emepdata}
\alias{findDuplicates.emepdata}
\title{Test an EMEP data file returned by the read.emepFile for duplicate
variables and parse back the indices of duplicate variables. The short
variable names, the matrices and the method_refs are tested for 
duplicates.}
\format{R Code}
\usage{
findDuplicates.emepdata(emepdata)
}
\arguments{
\item{emepdata}{; an EMEP data file given out by read.emepFile}
}
\value{
integer_array ; integer array ; an array of integers giving the  
   positions of duplicate variables in the input file (titleVarsShort,
   matrix and method_ref are tested)
}
\description{
WE HAVE TWO SIMILAR FUNCTIONS 'findDuplicatesNamesOnly.emepdata' AND 
'findDuplicates.emepdata'. WHILE THE FIRST ONE ONLY COMPARES THE short 
variable name, THE SECOND ONE COMPARES short variables name, matrix and 
method_ref.
}
\details{
The short variables names are emepdata$measurements$titleVarsShort
The matrices are emepdata$measurements$matrices
The method_refs are emepdata$measurements$method_refs
}
\note{
If no duplicates are present, an empty list is returned.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

