% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDuplicatesNamesOnly.emepdata.R
\name{findDuplicatesNamesOnly.emepdata}
\alias{findDuplicatesNamesOnly.emepdata}
\title{Test an EMEP data file returned by the read.emepFile for duplicate
variables and parse back the indices of duplicate variables. The short
variable names are tested for duplicates.}
\format{R Code}
\usage{
findDuplicatesNamesOnly.emepdata(emepdata)
}
\arguments{
\item{emepdata}{; an EMEP data file given out by read.emepFile}
}
\value{
integer_array ; integer array ; an array of integers giving the  
   positions of duplicate variables in the input file (titleVarsShort is 
   tested)
}
\description{
WE HAVE TWO SIMILAR FUNCTIONS 'findDuplicatesNamesOnly.emepdata' AND 
'findDuplicates.emepdata'. WHILE THE FIRST ONE ONLY COMPARES THE short 
variable name, THE SECOND ONE COMPARES short variables name, matrix and 
method_ref.
}
\details{
The short variables names are emepdata$measurements$titleVarsShort
}
\note{
If no duplicates are present, an empty list is returned.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

