% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProblematicAttribute.emepdata.R
\name{findProblematicAttributes.emepdata}
\alias{findProblematicAttributes.emepdata}
\title{Identify variables in EMEP data variable/list (emepdata) with problematic attributes}
\format{R Code}
\usage{
findProblematicAttributes.emepdata(emepdata)
}
\arguments{
\item{emepdata}{RODBC: an EMEP data file given out by read.emepFile}
}
\value{
\item{problemIndices}{integer array: indices of problematic variables}
 \item{allVars}{list(): a list containing one list per variable (not only of problematic ones); 
 the nested lists are numbered according to the variables in emepdata (not by variable name 
 because these might be duplicates). Each nested list has a 'name' and optionally a 'Fraction', 
 'Statistics' or further elements. Having other elements than 'name' does not indicate that
 the variable is problematic. Only the 'problemIndices' states, which variables are problematic.}
}
\description{
Identify variables in EMEP data variable/list (emepdata) with problematic attributes
}
\details{
Identify variables in EMEP data variable/list (emepdata) with problematic attributes

These are:
 - Fraction
 - Statistics && Statistics != 'arithmetic mean'
 - FURTHER???
 
These can be resolved:
 - Statistics: removeStatistics.emepdata(emepdata)
}
\note{
nothing
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

