% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmptyEmepData.R
\name{getEmptyEmepData}
\alias{getEmptyEmepData}
\title{Creates empty list according to read.emepFile .}
\format{R Code}
\usage{
getEmptyEmepData()
}
\value{
emep_file ; list ; Empty list which has (or should have) the same
   structure as the returned list from function read.emepFile.
}
\description{
Creates empty list according to read.emepFile .
}
\note{
Please update if read.emepFile is updated.
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

