% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxDoy.R
\name{getMaxDoy}
\alias{getMaxDoy}
\title{Calculate last day of a year as Julian date (366 or 365)}
\format{R Code}
\usage{
getMaxDoy(year)
}
\arguments{
\item{year}{(integer): The year (YYYY format) for which the last day of the year should be
calculated.}
}
\value{
doy (integer): The last day of the year "year" in the YYYYDDD format
}
\description{
Calculate the last day of the year "year" in the date format YYYYDDD. Year 
is an integer of the format YYYY. Thus, the year 1993 is parsed as 
year=1993 and NOT as year=93.
}
\details{
It is tested whether "year" is a leap year or not. If it is leap year
doy=YYYY366 and if not doy=YYYY365. doy always has 7 digits.

Alternative and shorter form that can be used directly in the code:

julian(as.Date(paste(year,"-12-31",sep="")), origin=as.Date(paste(year,"-01-01",sep="")))+1
}
\note{
There is an alternative and shorter form.
}
\examples{
  # choose a non-leap year (e.g. 1993)
  myYear = 1993
  myLastDoy = getMaxDoy(myYear)
  # 1993 is no leap year. Hence the output should be 1993365
  print(myLastDoy)
  
  
  # choose a leap year (e.g. 2004)
  myYear = 2004
  myLastDoy = getMaxDoy(myYear)
  # 2004 is a leap year. Hence the output should be 2004366
  print(myLastDoy)
  
  # an alternative method to calculate the last doy is:
  myYear = 2001
  (julian(as.Date(paste(myYear,"-12-31",sep="")), origin=as.Date(paste(myYear,"-01-01",sep="")))+1)[1]
  myYear = 2004
  (julian(as.Date(paste(myYear,"-12-31",sep="")), origin=as.Date(paste(myYear,"-01-01",sep="")))+1)[1]
  
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

