% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEntries.emepdata.R
\name{removeEntries.emepdata}
\alias{removeEntries.emepdata}
\title{Remove variables from an EMEP data variable/list (emepdata) as
parsed by removeList.}
\format{R Code}
\usage{
removeEntries.emepdata(emepdata, removeList = c())
}
\arguments{
\item{emepdata}{emepdata: an EMEP data file given out by read.emepFile}

\item{removeList}{integer array: (optional) an integer array containing
indices of variables which shoul be removed from emepdata;}
}
\value{
new_data a list() of the same format as that 
         given back by the read.emepFile but without duplicate 
         variables (titleVarsShort)
}
\description{
Remove variables from an EMEP data variable/list (emepdata) as
parsed by removeList.
}
\details{
Following parts of emepdata are changed if duplicates are present:
 drop elements:
   measurements$titleVarsLong
   measurements$titleVarsShort
   measurements$units
   measurements$matrices
   measurements$method_refs
   measurements$variables
     note: variables is a 2-dim array; the 2nd dimension has to be changed
   flags$titleFlagsLong
   flags$titleFlagsShort
   flags$numflags
     note: variables is a 2-dim array; the 2nd dimension has to be changed
 alter integer:
   nVars
}
\note{
If no duplicates are present, new_data equals emepdata and changes
   is FALSE
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

