% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeStatistics.emepdata.R
\name{removeStatistics.emepdata}
\alias{removeStatistics.emepdata}
\title{Remove all statistic variables from an EMEP data variable/list (emepdata),
which are unequal 'arithmetic mean'.}
\format{R Code}
\usage{
removeStatistics.emepdata(emepdata)
}
\arguments{
\item{emepdata}{emepdata: an EMEP data file given out by read.emepFile}
}
\value{
\item{changes}{logical: FALSE if new_data equals
         emepdata and TRUE otherwise (= if variables were dropped)}
 \item{removeList}{integer array: the same integer array as returned
         by findDuplicates.emepdata(emepdata)}
 \item{new_data}{list(): list of the same format as that 
         given back by the read.emepFile but without duplicate 
         variables (titleVarsShort)}
}
\description{
Remove all statistic variables from an EMEP data variable/list (emepdata),
which are unequal 'arithmetic mean'.
}
\details{
Remove all statistic variables from an EMEP data variable/list (emepdata),
which are unequal 'arithmetic mean'.

The function removeEntries.emepdata removes the entries.
}
\note{
no note
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

