% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reworkTiming.R
\name{reworkTiming}
\alias{reworkTiming}
\title{Convert a time}
\format{R Code}
\usage{
reworkTiming(times, refDate)
}
\arguments{
\item{times}{;}

\item{refDate}{;}
}
\value{
newTimes ;
}
\description{
The array times contains dates of measurements in
days (real variable; + hours divided by 24) with 
respect to a reference date refDate. refDate has to
be given in Julian format (YYYYDDD).
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

