% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifyStationId.R
\name{verifyStationId}
\alias{verifyStationId}
\title{Tests whether a station_id has the correct format}
\format{R Code}
\usage{
verifyStationId(station_id)
}
\arguments{
\item{station_id}{character: id of an measurement station}
}
\value{
list(station_id, err_code, warn_code) list(character, integer): err_code (0 = success; 1 err), warn_code (0 = no warn; 1 = warning);
}
\description{
Tests whether a station_id has the correct format
}
\author{
Daniel Neumann, Helmholtz-Zentrum Geesthacht
   \email{daniel.neumann@___.de}
}

